refactor(githubactions-python):  be consistent with the plugin template
