--- internal/pkg/plugin/githubactions/python/helper.go
@@ -1,26 +1,0 @@
-package python
-
-import (
-	"fmt"
-
-	"github.com/mitchellh/mapstructure"
-
-	"github.com/merico-dev/stream/pkg/util/log"
-)
-
-func parseAndValidateOptions(options map[string]interface{}) (*Options, error) {
-	var opts Options
-	err := mapstructure.Decode(options, &opts)
-	if err != nil {
-		return nil, err
-	}
-
-	if errs := validate(&opts); len(errs) != 0 {
-		for _, e := range errs {
-			log.Errorf("Options error: %s.", e)
-		}
-		return nil, fmt.Errorf("opts are illegal")
-	}
-
-	return &opts, nil
-}
